Register unseen vue files when TSLS requests them
The Typescript language service references files that are not open for
editing, and thus don't have entries in the language model cache. Since
they don't have entries, there is no way to know whether their script
block is Javascript or Typescript.

The fix is to load these files and register them with the language model cache.