End vue-html grammar only at beginning of line
End the vue-html only with a `</template>` at the beginning of a line.
Because the `<template>` tag starts a grammar in source.vue files and
can also be used a a tag *within* that grammar, the closing
`</template>` of the grammar must be distinguishable from the same
closing tag within the vue-html grammar.

This enables removing the recursive application of the vue-html grammar
within the vue-html grammar because source.vue can reliably know when
a `</template>` should end the grammar.

* New restriction: there can be only 1 closing `</template>` within a
  Vue single-file component (SFC) otherwise the vue-html grammar will
  be closed early. That closing `</template>` must also appear at the
  beginning of a line. Because this follows Vue formatting guidelines,
  this should have zero impact on users who following those guidelines.

Fixes #1211