Refactor isVX functions to remove type predicates
Type guard functions that use type predicates only do one thing:
They confirm that the type of the parameter is as specified in the
predicate.

However, functions like `isVBind` were also checking string
properties to check the content of the object structure, which
made them do more than just confirm that the type is
VDirective. This meant that the functions were being used
as conditionals for more than just their type guarding, giving
unexpected results.

The solution is to create a function, `isVDirective`, which is the
type guard for `node is AST.VDirective`, and change the returns
of the other `isVX` functions to be `boolean`.