fix(interpolation/v-for): allow optional index
The `v-for` directive when value is a number provides both the 1-based numeric value, and the 0-based index to the iteration function, however vetur currently shows this as an error when you are using the experimental.templateInterpolationService option.

This change adds the missing second parameter to the iteration helper for the `num: number` overload, and existing usages with a single parameter should continue to work.